conversion vers unsigned long long int echec, a corriger

+++ Imperative2/tp4/direct.h
@@ -2,6 +2,6 @@
 #include <stdio.h>
 #include <stdlib.h>
 
-int fibiter (int);
-int lire (FILE *, int);
-int remplir (FILE *, int);
+unsigned long long int fibiter (int);
+unsigned long long int  lire (FILE *, int);
+unsigned long long int  remplir (FILE *, int);
